fix(install): move #200 helpers to install-helpers.cjs so tests don't run installer (#200)
The new test required bin/install.js to reach the pure helpers, but install.js
runs the installer on require (non-interactive auto-install) — which fails the
test file at load under the full CI runner (the local isolated run masked it).

- Move shouldCopyToNfBin / isUnderInstallDir / synthesizeMcpEntry /
  installedUnifiedMcpPath / NF_BIN_RUNTIME_MJS into the existing pure
  install-helpers.cjs (already "extracted from install.js for unit testability").
- install.js imports them from there (still re-exports for compat).
- Test imports from install-helpers.cjs — no installer side effect.
